Restructure code a bit
