+++ PluginsAndFeatures/azure-toolkit-for-intellij/gradle.properties
@@ -1,4 +1,4 @@
-buildNumber=3.46.0.9999-2020.3
+buildNumber=3.46.0.9999-2021.1
 
 javaVersion=11
 org.gradle.jvmargs='-Duser.language=en'
@@ -17,7 +17,7 @@ updateVersionRange=false
 patchPluginXmlSinceBuild=211.4961.4
 
 kotlin_version=1.4.20
-jackson_kotlin_version=2.11.0
+jackson_kotlin_version=2.12.0
 gradle_plugin_version=0.6.5
 undercouch_version=4.0.4
 web_socket_version=1.5.1
