+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/ReSharper.Azure/src/Azure.Intellisense/FunctionApp/LiveTemplates/Scope/AzureFSharpProjectScopeCategoryUIProvider.cs
@@ -1,0 +1,42 @@
+﻿// Copyright (c) 2021 JetBrains s.r.o.
+// <p/>
+// All rights reserved.
+// <p/>
+// MIT License
+// <p/>
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+// documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
+// to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+// <p/>
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+// the Software.
+// <p/>
+// THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+// THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using System.Collections.Generic;
+using JetBrains.ProjectModel.Resources;
+using JetBrains.ReSharper.Feature.Services.LiveTemplates.Scope;
+
+namespace JetBrains.ReSharper.Azure.Intellisense.FunctionApp.LiveTemplates.Scope
+{
+    [ScopeCategoryUIProvider(Priority = -41.0, ScopeFilter = ScopeFilter.Project)]
+    public class AzureFSharpProjectScopeCategoryUIProvider : ScopeCategoryUIProvider
+    {
+        public AzureFSharpProjectScopeCategoryUIProvider() : base(ProjectModelThemedIcons.Fsharp.Id)
+        {
+            MainPoint = new InAzureFunctionsFSharpProject();
+        }
+
+        public override IEnumerable<ITemplateScopePoint> BuildAllPoints()
+        {
+            yield return new InAzureFunctionsFSharpProject();
+        }
+
+        public override string CategoryCaption => "Azure (F#)";
+    }
+}
